requirements.txt
PySide6>=6.7
PySide6-Addons>=6.7
PySide6-Essentials>=6.7
requests>=2.32
PyJWT>=2.8
python-dotenv>=1.0
